Generation of note chain given a map
using the beat saver hashes in maplist, running generation.py will create a list of each note chain that appears. It is sorted from most often to least often. SLIDER_TIMING is a global variable that determines how much time in front of a note to ignore in order to correct for sliders.